Fix post merge serializer failure
